Updating Kiwi

+++ Kiwi/KWMessageTracker.m
@@ -96,12 +96,12 @@
     if (aCount == 1)
         return @"1 time";
 
-    return [NSString stringWithFormat:@"%d times", aCount];
+    return [NSString stringWithFormat:@"%d times", (int)aCount];
 }
 
 - (NSString *)expectedCountPhrase {
     NSString *countPhrase = [self phraseForCount:self.count];
-    
+
     switch (self.countType) {
         case KWCountTypeExact:
             return [NSString stringWithFormat:@"exactly %@", countPhrase];
@@ -144,8 +144,8 @@
     return [NSString stringWithFormat:@"messagePattern: %@\nmode: %@\ncount: %d\nreceiveCount: %d",
                                       self.messagePattern,
                                       self.modeString,
-                                      self.count,
-                                      self.receivedCount];
+                                      (int)self.count,
+                                      (int)self.receivedCount];
 }
 
 @end
